# src/static/swagger.json
{
    "info": {
        "description": "Auth service",
        "title": "Auth service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0441\u043f\u0438\u0441\u043a\u0430 \u0432\u0441\u0435\u0445 \u0443\u0440\u043e\u0432\u043d\u0435\u0439 \u0434\u043e\u0441\u0442\u0443\u043f\u0430 \u0438\u0437 \u0431\u0430\u0437\u044b",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "204": {
                        "description": "Permissions list is empty"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "tags": [
                    "Permission"
                ]
            }
        },
        "/api/v1/permissions/{role_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0441\u043f\u0438\u0441\u043a\u0430 \u0432\u0441\u0435\u0445 \u0443\u0440\u043e\u0432\u043d\u0435\u0439 \u0434\u043e\u0441\u0442\u0443\u043f\u0430 \u0440\u043e\u043b\u0438",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "204": {
                        "description": "Permissions list is empty"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "tags": [
                    "Permission"
                ]
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u0414\u043e\u0431\u0430\u0432\u043b\u0435\u043d\u0438\u0435 \u0443\u0440\u043e\u0432\u043d\u044f \u0434\u043e\u0441\u0442\u0443\u043f\u0430 \u043a \u0440\u043e\u043b\u0438",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "permission_id",
                        "in": "query",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission assigned to role"
                    },
                    "204": {
                        "description": "Permissions list is empty"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Permission is already in use"
                    }
                },
                "tags": [
                    "Permission"
                ]
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u0423\u0434\u0430\u043b\u0435\u043d\u0438\u0435 \u0443\u0440\u043e\u0432\u043d\u044f \u0434\u043e\u0441\u0442\u0443\u043f\u0430 \u0443 \u0440\u043e\u043b\u0438",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "permission_id",
                        "in": "query",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "204": {
                        "description": "Permissions list is empty"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Permission is already deleted"
                    }
                },
                "tags": [
                    "Permission"
                ]
            }
        },
        "/api/v1/permission": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u0414\u043e\u0431\u0430\u0432\u043b\u0435\u043d\u0438\u0435 \u043d\u043e\u0432\u043e\u0433\u043e \u0443\u0440\u043e\u0432\u043d\u044f \u0434\u043e\u0441\u0442\u0443\u043f\u0430",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Permission"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New permission was created"
                    },
                    "409": {
                        "description": "Permission is already in use"
                    },
                    "403": {
                        "description": "Permission denied"
                    }
                },
                "tags": [
                    "Permission"
                ]
            }
        },
        "/api/v1/permission/{permission_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u041f\u0440\u043e\u0441\u043c\u043e\u0442\u0440 \u0443\u0440\u043e\u0432\u043d\u044f \u0434\u043e\u0441\u0442\u0443\u043f\u0430 \u043f\u043e id",
                "parameters": [
                    {
                        "name": "permission_id",
                        "in": "path",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "204": {
                        "description": "Role list is empty"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "tags": [
                    "Permission"
                ]
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u0418\u0437\u043c\u0435\u043d\u0438\u0435 \u0443\u0440\u043e\u0432\u043d\u044f \u0434\u043e\u0441\u0442\u0443\u043f\u0430 \u043f\u043e id",
                "parameters": [
                    {
                        "name": "permission_id",
                        "in": "path",
                        "type": "string",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Permission"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission was changed sucessfully"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "tags": [
                    "Permission"
                ]
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "\u0423\u0434\u0430\u043b\u0435\u043d\u0438\u0435 \u0443\u0440\u043e\u0432\u043d\u044f \u0434\u043e\u0441\u0442\u0443\u043f\u0430 \u043f\u043e id",
                "parameters": [
                    {
                        "name": "permission_id",
                        "in": "path",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "tags": [
                    "Permission"
                ]
            }
        },
...
    },
...
    "openapi": "3.0.2",
    "components": {
        "schemas": {
...
            "Permission": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "description": "code",
                        "example": 0
                    },
                    "description": {
                        "type": "string",
                        "description": "description",
                        "example": "\u041c\u043e\u0436\u0435\u0442 \u043f\u0440\u043e\u0441\u043c\u0430\u0442\u0440\u0438\u0432\u0430\u0442\u044c \u0441\u0432\u043e\u0439 \u043a\u043e\u043d\u0442\u0435\u043d\u0442"
                    },
                    "name": {
                        "type": "string",
                        "description": "name",
                        "example": "Default user"
                    }
                },
                "required": [
                    "code",
                    "description",
                    "name"
                ]
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}